.gitignore:
Added etc/ to the list of ignored files and directories.
Dockerfile:

Uncommented the line #ADD ./ArchiveBox.conf ./ArchiveBox.conf to add the ArchiveBox.conf file to the Docker image.
docker-compose.yml:

Added container_name: archivebox to the archivebox service.
Added the following port mappings to the pihole service: 53/tcp, 53/udp, and 67/udp.
Added the WEBPASSWORD environment variable to the pihole service, with the value adminpassword.
Added the volumes section to the pihole service.